Rating/requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
export==0.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.15.7
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
